LB and UB as float64 to easily compare with a fitness

+++ pkg/pfsp/pfsp.go
@@ -6,8 +6,8 @@ type Instance struct {
 	Jobs       int
 	Machines   int
 	Seed       int
-	UpperBound int
-	LowerBound int
+	UpperBound float64
+	LowerBound float64
 	Instance   int
 	Matrix     [][]float64
 }
